restaurants_project/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Deprecated==1.2.13
distro==1.7.0
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
environ==1.0
idna==3.3
iniconfig==1.1.1
jsonschema==3.2.0
kombu==5.2.4
packaging==21.3
paramiko==2.11.0
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.14.1
